ensure nstn is an int in StaticUtils

+++ SimPEG/EM/Static/Utils/StaticUtils.py
@@ -207,7 +207,7 @@ def gen_DCIPsurvey(endl, mesh, surveyType, a, b, n):
     dl_x = (endl[1, 0] - endl[0, 0]) / dl_len
     dl_y = (endl[1, 1] - endl[0, 1]) / dl_len
 
-    nstn = np.floor(dl_len / a)
+    nstn = int(np.floor(dl_len / a))
 
     # Compute discrete pole location along line
     stn_x = endl[0, 0] + np.array(range(int(nstn)))*dl_x*a
